cac_requirements.txt
absl-py==0.9.0
addict==2.4.0
alabaster==0.7.12
albumentations==1.0.3
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
attrs==21.4.0
Babel==2.10.3
black==21.4b2
bleach==5.0.0
botorch==0.2.0
cachetools==4.1.0
certifi==2022.6.15
cffi==1.15.0
cfgv==3.3.1
chardet==3.0.4
cityscapesScripts @ git+https://github.com/mcordts/cityscapesScripts.git@aeb7b82531f86185ce287705be28f452ba3ddbb8
click==7.1.2
click-default-group==1.2.2
cloudpickle==2.1.0
colorama==0.4.5
coloredlogs==15.0.1
commonmark==0.9.1
ConfigSpace==0.4.18
contextlib2==21.6.0
cryptography==37.0.2
cycler==0.11.0
Cython==0.29.21
detail==4.0
detectron2==0.6+cu111
distlib==0.3.4
docutils==0.16
einops==0.5.0
et-xmlfile==1.1.0
filelock==3.7.1
flake8==4.0.1
Flask==2.0.1
formaldict==1.0.5
future==0.18.2
fvcore==0.1.5.post20220512
google-auth==1.17.0
google-auth-oauthlib==0.4.1
gpytorch==1.4.2
grpcio==1.29.0
h5py==3.7.0
humanfriendly==10.0
hydra-core==1.2.0
hyperopt==0.2.5
identify==2.5.1
idna==2.9
imagecorruptions==1.1.2
imageio==2.9.0
imagesize==1.3.0
importlib-metadata==4.11.4
importlib-resources==5.8.0
iniconfig==1.1.1
iopath==0.1.9
itsdangerous==2.1.2
jeepney==0.8.0
Jinja2==3.1.2
joblib==1.1.0
keyring==23.6.0
kiwisolver==1.4.3
kmatch==0.5.0
Markdown==3.2.2
MarkupSafe==2.1.1
matplotlib==3.3.4
mccabe==0.6.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mmcls==0.25.0
mmcv-full==1.5.0
mmpycocotools==12.0.3
mmsegmentation==0.30.0
mock==4.0.3
MultiScaleDeformableAttention==1.0
mypy-extensions==0.4.3
networkx==2.8.4
nms==0.1.6
nodeenv==1.6.0
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1652801679809/work
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
omegaconf==2.2.2
onnx==1.11.0
onnxruntime==1.3.0
opencv-python==4.6.0.66
opencv-python-headless==4.2.0.34
openpyxl==3.0.7
packaging==21.3
pandas==1.1.5
panopticapi @ git+https://github.com/cocodataset/panopticapi.git@7bb4655548f98f3fedc07bf37e9040a992b054b0
pathspec==0.9.0
patsy==0.5.2
Pillow==7.1.2
pkginfo==1.8.3
platformdirs==2.5.2
pluggy==1.0.0
Polygon3==3.0.9.1
portalocker==2.4.0
pre-commit==2.14.0
prettytable==3.5.0
prompt-toolkit==3.0.32
protobuf==3.20.0
ptflops==0.6.4
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0
pycocotools==2.0.4
pycodestyle==2.8.0
pycparser==2.21
pydot==1.4.2
pyflakes==2.4.0
Pygments==2.12.0
pyparsing==2.4.7
pyquaternion==0.9.9
pytest==7.1.2
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==5.3.1
readme-renderer==35.0
recommonmark==0.6.0
regex==2022.7.9
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==2.0.0
rsa==4.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.6
schema==0.7.4
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.5.3
SecretStorage==3.3.2
Shapely==1.7.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
SMore-core-full==2.0.5
SMore-seg==2.0.2
snowballstemmer==2.2.0
Sphinx==3.2.0
sphinx-markdown-tables==0.0.15
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
statsmodels==0.12.2
submitit==1.4.2
tabulate==0.8.9
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2022.5.4
timm==0.5.4
toml==0.10.2
tomli==2.0.1
torch==1.9.1
torchaudio==0.9.0a0+a85b239
torchvision==0.10.1
tqdm==4.54.1
twine==3.3.0
typing==3.7.4.3
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1656706066251/work
urllib3==1.25.9
virtualenv==20.14.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
wget==3.2
yacs==0.1.8
yapf==0.32.0
zipp==3.1.0
